добавил передачу координат

+++ map/models.py
@@ -27,8 +27,8 @@ class UnverifiedTag(models.Model):
     description = models.TextField('Описание', max_length=100)
     image = models.ImageField('Изображение', upload_to='images/')
     location = models.CharField('Местоположение', max_length=30)
-    x_coord = models.PositiveIntegerField('Координата X')
-    y_coord = models.PositiveIntegerField('Координата Y')
+    x_coord = models.FloatField('Координата X')
+    y_coord = models.FloatField('Координата Y')
     username = models.CharField('Имя пользователя', max_length=30)
     email = models.EmailField('Email', max_length=30)
